requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
etils==1.12.2
executing==2.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpath-rw==1.4.0
jsonpointer==3.0.0
kagglehub==0.3.12
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-ollama==0.3.3
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mlcroissant==1.0.17
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
ollama==0.4.8
openai==1.82.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-cli-ui==0.7.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
types-pytz==2025.2.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0